LayoutTests/media/modern-media-controls/fullscreen-support/fullscreen-support-enabled-expected.txt

Testing the FullscreenSupport behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.fullscreenButton.enabled is true
PASS mediaController.controls.fullscreenButton.enabled is false

PASS successfullyParsed is true

TEST COMPLETE
